'analyze' module now has tests.
As always, had to be refactored to allow for meaningful testing. The
closer you get to functional programming, the easier it is for write
meaningful tests - though I can't say clearly that oo is hard to test in
itself, but it definitely encourages types of coding which are hard to
test.