Issue #24
Change names of the 'test' and 'stats' indices. Allowing for more
flexible configuration - the name of the index into which test data is
loaded is now controlled by the config.json file ('name_index' field)
and the name of the 'stats' index is now determined by the
esbench.STATS_INDEX_NAME global variable.