Initial work on issue/34
The issue is to allow specifying not only the number of documents to be
uploaded, but also the byte size of the documents. This is more
meaningful in terms of performance measurements.

The big architectural change is on how data is being fed into the
benchmarker - now it is an iterator of iterators - an iterator which
yields batches, which in turn yield individual documents.