Better logging around sizes / counts of inserts.
Accurate counts / sizes of incoming data now logged. Turns out that the
size of the index is not the same as the size of the raw data that is
being inserted - the precise relationship is yet to be established, for
now the logging at least gives an idea of what is going on. I'd go out
on a limb and say that this closes #34.