Ability to flatten benchmark and observation data.
This is for #54; benchmark and observation data are complex
dictionaries. esbench.analyze.flat() function will combine benchmark and
observation data for each observation, and flatten the data into a list
of (name, value) tuples, where name is built by combining all the
dictionary key names leaving to the value, joining keys with periods. If
there are lists, element indexes are inserted, so you may end up with a
tuple like: ('foo.bar.1.x', 'baz'). flat() also takes a list of regex
strings which will be used to filter the resulting key strings.