Updated bench to work with multiple shards.
Segment stats are now correctly aggregated in Observation._segments()
when dealing with an index which has multiple shards, primary [and]
secondary.

In addition, ability to specify the number of shards (primaries) at
command line with the '--shards' argument.

Ability to record detailed per-segment information removed (no need for
that kind of detail, leads to needlessly complex observation data).

Version bump to 0.2.1